Acquisition&Messaging/py-requirements.txt
kafka-python-ng==2.2.2
openmeteo-requests==1.3.0
requests-cache==1.2.1
retry-requests==2.0.0